requirement.txt
aerich==0.9.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asyncclick==8.2.2.2
asyncpg==0.30.0
bcrypt==4.3.0
click==8.2.1
dictdiffer==0.9.0
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
iso8601==2.1.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pypika-tortoise==0.6.1
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0